[Refactor] - Abstract out components part1
